refactor: adding deep-speaker sub-module
