+++ speaker_verification/deep_speaker/rescnn_model.py
@@ -21,7 +21,7 @@ from tensorflow.keras.layers import (
 )
 from tensorflow.keras.models import Model
 
-from speaker_verification.audio import NUM_FBANKS, NUM_FRAMES
+from speaker_verification.deep_speaker.audio import NUM_FBANKS, NUM_FRAMES
 
 logger = logging.getLogger(__name__)
